requirements.txt
fastapi==0.128.0
uvicorn==0.31.0
pymupdf
pydantic==2.12.5
python-dotenv==1.2.1
requests==2.32.5
# LangChain embeddings with HuggingFace
langchain==0.1.20
langchain-huggingface==0.0.16
sentence-transformers==3.0.1
# Commented out: Supabase and database connections
# psycopg2==2.9.11
# supabase==2.27.2
# tqdm==4.67.1